double -> float
migration to float values instead of double

+++ core/src/net/pzdcrp/Aselia/data/DM.java
@@ -1,10 +1,10 @@
 package net.pzdcrp.Aselia.data;
 
 public class DM {
-	public static final double badVel = 0.005;
-	public static final double walkSpeed = 0.03*3;
-	public static final double runSpeed = 0.05*3;
-	public static final double gravity = 0.08;
-	public final static double airdrag = 0.9800000190734863D;
+	public static final float badVel = 0.005f;
+	public static final float walkSpeed = 0.03f*3f;
+	public static final float runSpeed = 0.05f*3f;
+	public static final float gravity = 0.08f;
+	public final static float airdrag = 0.9800000190734863f;
 	public static final int maxll = 7;
 }
